+++ src/geometryRTheta/poisson/poisson_like_rhs_function.hpp
@@ -2,8 +2,10 @@
 #pragma once
 
 #include <ddc/ddc.hpp>
+#include <ddc/kernels/splines.hpp>
 
 #include "geometry.hpp"
+#include "spline_definitions.hpp"