+++ src/geometryXY/geometry/geometry.hpp
@@ -1,8 +1,6 @@
 // SPDX-License-Identifier: MIT
-
 #pragma once
 #include <ddc/ddc.hpp>
-#include <ddc/kernels/splines.hpp>
 
 #include "ddc_alias_inline_functions.hpp"
 #include "ddc_aliases.hpp"
@@ -58,87 +56,14 @@ using CoordX = Coord<X>;
 using CoordY = Coord<Y>;
 using CoordXY = Coord<X, Y>;
 
-int constexpr BSDegreeX = 3;
-int constexpr BSDegreeY = 3;
-
-bool constexpr BsplineOnUniformCellsX = true;
-bool constexpr BsplineOnUniformCellsY = true;
-
-struct BSplinesX
-    : std::conditional_t<
-              BsplineOnUniformCellsX,
-              ddc::UniformBSplines<X, BSDegreeX>,
-              ddc::NonUniformBSplines<X, BSDegreeX>>
-{
-};
-struct BSplinesY
-    : std::conditional_t<
-              BsplineOnUniformCellsY,
-              ddc::UniformBSplines<Y, BSDegreeY>,
-              ddc::NonUniformBSplines<Y, BSDegreeY>>
-{
-};
-
-ddc::BoundCond constexpr SplineXBoundary = ddc::BoundCond::PERIODIC;
-ddc::BoundCond constexpr SplineYBoundary = ddc::BoundCond::PERIODIC;
-
-// IDim initialisers
-using SplineInterpPointsX
-        = ddc::GrevilleInterpolationPoints<BSplinesX, SplineXBoundary, SplineXBoundary>;
-using SplineInterpPointsY
-        = ddc::GrevilleInterpolationPoints<BSplinesY, SplineYBoundary, SplineYBoundary>;
-
 // IDim definitions
-struct GridX : SplineInterpPointsX::interpolation_discrete_dimension_type
+struct GridX : UniformGridBase<X>
 {
 };
-struct GridY : SplineInterpPointsY::interpolation_discrete_dimension_type
+struct GridY : UniformGridBase<Y>
 {
 };
 
-
-// SplineBuilder and SplineEvaluator definitions
-using SplineXBuilder_XY = ddc::SplineBuilder<
-        Kokkos::DefaultExecutionSpace,
-        Kokkos::DefaultExecutionSpace::memory_space,
-        BSplinesX,
-        GridX,
-        SplineXBoundary,
-        SplineXBoundary,
-        ddc::SplineSolver::LAPACK>;
-using SplineXEvaluator_XY = ddc::SplineEvaluator<
-        Kokkos::DefaultExecutionSpace,
-        Kokkos::DefaultExecutionSpace::memory_space,
-        BSplinesX,
-        GridX,
-        ddc::PeriodicExtrapolationRule<X>,
-        ddc::PeriodicExtrapolationRule<X>>;
-
-
-using SplineYBuilder_XY = ddc::SplineBuilder<
-        Kokkos::DefaultExecutionSpace,
-        Kokkos::DefaultExecutionSpace::memory_space,
-        BSplinesY,
-        GridY,
-        SplineYBoundary,
-        SplineYBoundary,
-        ddc::SplineSolver::LAPACK>;
-using SplineYEvaluator_XY = ddc::SplineEvaluator<
-        Kokkos::DefaultExecutionSpace,
-        Kokkos::DefaultExecutionSpace::memory_space,
-        BSplinesY,
-        GridY,
-        ddc::PeriodicExtrapolationRule<Y>,
-        ddc::PeriodicExtrapolationRule<Y>>;
-
-// Spline index range
-using IdxRangeBSX = IdxRange<BSplinesX>;
-using IdxRangeBSY = IdxRange<BSplinesY>;
-using IdxRangeBSXY = IdxRange<BSplinesX, BSplinesY>;
-
-template <class ElementType>
-using BSConstFieldXY = Field<ElementType const, IdxRangeBSXY>;
-using DBSConstFieldXY = BSConstFieldXY<double>;
 
 // Index definitions
 using IdxX = Idx<GridX>;