+++ tests/geometryRTheta/quadrature/tests_L1_and_L2_norms.cpp
@@ -10,6 +10,7 @@
 #include "l_norm_tools.hpp"
 #include "mesh_builder.hpp"
 #include "quadrature.hpp"
+#include "spline_definitions.hpp"
 #include "spline_quadrature.hpp"
 #include "trapezoid_quadrature.hpp"
 #include "volume_quadrature_nd.hpp"