+++ tests/geometryXVx/spatialadvection.cpp
@@ -4,6 +4,7 @@
 
 #include "bsl_advection_x.hpp"
 #include "geometry.hpp"
+#include "spline_definitions.hpp"
 #include "spline_interpolator.hpp"